apps/api-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.8
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.1
aiofiles==23.2.1
httpx==0.25.2

# For future async jobs
# celery==5.3.4
# redis==5.0.1

# For future database
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# For development
pytest==7.4.3
pytest-asyncio==0.21.1